// apps/transactions/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const TxnType = {
  TOPUP: 'TOPUP',
  TRANSFER: 'TRANSFER',
  RECURRING_TOPUP: 'RECURRING_TOPUP',
  RECURRING_TRANSFER: 'RECURRING_TRANSFER',
  SPLIT_PAYMENT: 'SPLIT_PAYMENT',
  REQUEST_PAYMENT: 'REQUEST_PAYMENT'
} as const

export type TxnType = (typeof TxnType)[keyof typeof TxnType]


export const TxnStatus = {
  PENDING: 'PENDING',
  PROCESSING: 'PROCESSING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
  CANCELLED: 'CANCELLED',
  REVERSED: 'REVERSED'
} as const

export type TxnStatus = (typeof TxnStatus)[keyof typeof TxnStatus]


export const Category = {
  FOOD: 'FOOD',
  TRANSPORT: 'TRANSPORT',
  ENTERTAINMENT: 'ENTERTAINMENT',
  TRANSFER: 'TRANSFER',
  UTILITIES: 'UTILITIES',
  OTHER: 'OTHER'
} as const

export type Category = (typeof Category)[keyof typeof Category]